Revert "Move spirv_generator into _kernel_api_impl.spirv"
This reverts commit 779b09636199e8d37f9afe2937accbe658e0f7e9.
